requirements.txt
alembic==1.10.2
aniso8601==9.0.1
astroid==2.15.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorclass==2.2.2
coverage==6.5.0
coveralls==3.3.1
cryptography==39.0.2
dill==0.3.6
dnspython==2.3.0
docopt==0.6.2
email-validator==1.3.1
exceptiongroup==1.1.0
Flask==2.2.3
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mysql-connector-python==8.0.32
mysql-connector-python-rf==2.2.2
packaging==23.0
pip-upgrader==1.4.15
platformdirs==3.1.0
pluggy==1.0.0
protobuf==3.20.3
pycparser==2.21
pylint==2.17.0
PyMySQL==1.0.2
pytest==7.2.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.5
terminaltables==3.1.10
tomli==2.0.1
tomlkit==0.11.6
gunicorn==20.1.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
